Ensure Hive box is opened asynchronously and handle default language
Updated the `LanguageLocalDataSourceImpl` to open the Hive box asynchronously in both `updateLanguage` and `getPreferredLanguage` methods. Also, modified `getPreferredLanguage` to return a default language ('en') if no preferred language is found. Additionally, fixed the route generation in `MovieAppState` to handle undefined routes gracefully by defaulting to the initial route. Lastly, updated the `MNavigationDrawer` to use `MaterialPageRoute` for navigating to the `FavoriteScreen` and ensured the `FavoriteScreen` is imported where necessary.